+++ WorldBuilder/Modules/Landscape/ViewModels/EnvCellStaticObjectViewModel.cs
@@ -15,7 +15,7 @@ public partial class EnvCellStaticObjectViewModel : SelectedObjectViewModelBase 
 
     public string ObjectIdHex => $"0x{ObjectId:X8}";
 
-    public EnvCellStaticObjectViewModel(uint objectId, ulong instanceId, uint landblockId, uint cellId, Vector3 position, Vector3 localPosition, Quaternion rotation) 
+    public EnvCellStaticObjectViewModel(uint objectId, ulong instanceId, ushort landblockId, uint cellId, Vector3 position, Vector3 localPosition, Quaternion rotation) 
         : base(instanceId, landblockId, position, localPosition, rotation) {
         ObjectIdVal = objectId;
         CellId = cellId;
